+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Holidays for Australia
 - Holidays for Bulgaria
 
 ## [0.9.0] - 2024-06-01